+++ front/pages/api/w/[wId]/mcp/index.ts
@@ -4,7 +4,7 @@ import type { NextApiRequest, NextApiResponse } from "next";
 
 import { internalMCPServerNameToSId } from "@app/lib/actions/mcp_helper";
 import { isInternalMCPServerName } from "@app/lib/actions/mcp_internal_actions/constants";
-import { fetchRemoteServerMetaDataByURL } from "@app/lib/actions/mcp_metadata";
+import { fetchRemoteMCPServerByURL } from "@app/lib/actions/mcp_metadata";
 import { withSessionAuthenticationForWorkspace } from "@app/lib/api/auth_wrappers";
 import type { MCPServerType, MCPServerTypeWithViews } from "@app/lib/api/mcp";
 import type { Authenticator } from "@app/lib/auth";
@@ -119,7 +119,7 @@ async function handler(
           });
         }
 
-        const metadata = await fetchRemoteServerMetaDataByURL(auth, url);
+        const metadata = await fetchRemoteMCPServerByURL(auth, url);
 
         const newRemoteMCPServer = await RemoteMCPServerResource.makeNew(auth, {
           workspaceId: auth.getNonNullableWorkspace().id,
